Move basis_status.{cpp,h} to utils, fix linking issues
