+++ src/solver/optimisation/CMakeLists.txt
@@ -8,9 +8,7 @@ set(RTESOLVER_OPT
         opt_pilotage_optimisation_lineaire.cpp
         opt_pilotage_optimisation_quadratique.cpp
 		include/antares/solver/optimisation/opt_structure_probleme_a_resoudre.h
-		include/antares/solver/optimisation/basis_status.h
 		include/antares/solver/optimisation/opt_constants.h
-	basis_status.cpp
 	opt_structure_probleme_a_resoudre.cpp
         opt_alloc_probleme_a_optimiser.cpp
         opt_gestion_des_bornes_cas_quadratique.cpp
@@ -196,7 +194,7 @@ else ()
 endif ()
 target_link_libraries(model_antares
         PUBLIC
-        utils
+        Antares::solverUtils
         sirius_solver
         antares-solver-simulation
         Antares::benchmarking
