+++ src/solver/utils/CMakeLists.txt
@@ -11,6 +11,8 @@ set(SRC
         name_translator.cpp
         include/antares/solver/utils/opt_period_string_generator.h
         opt_period_string_generator.cpp
+	basis_status.cpp
+	include/antares/solver/utils/basis_status.h
 )
 
 add_library(utils ${SRC})
