+++ src/solver/utils/named_problem.cpp
@@ -19,7 +19,7 @@
 ** along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
 */
 #include "antares/solver/utils/named_problem.h"
-#include "antares/solver/optimisation/basis_status.h"
+#include "antares/solver/utils/basis_status.h"
 #include <algorithm>
 
 namespace Antares::Optimization
